+++ saleor/graphql/attribute/schema.py
@@ -5,14 +5,10 @@ from ..translations.mutations import AttributeTranslate, AttributeValueTranslate
 from .bulk_mutations import AttributeBulkDelete, AttributeValueBulkDelete
 from .filters import AttributeFilterInput
 from .mutations import (
-    AttributeClearMeta,
-    AttributeClearPrivateMeta,
     AttributeCreate,
     AttributeDelete,
     AttributeReorderValues,
     AttributeUpdate,
-    AttributeUpdateMeta,
-    AttributeUpdatePrivateMeta,
     AttributeValueCreate,
     AttributeValueDelete,
     AttributeValueUpdate,
@@ -52,32 +48,6 @@ class AttributeMutations(graphene.ObjectType):
     attribute_translate = AttributeTranslate.Field()
     attribute_bulk_delete = AttributeBulkDelete.Field()
     attribute_value_bulk_delete = AttributeValueBulkDelete.Field()
-
-    # meta mutations
-    attribute_update_metadata = AttributeUpdateMeta.Field(
-        deprecation_reason=(
-            "Use the `updateMetadata` mutation instead. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-    attribute_clear_metadata = AttributeClearMeta.Field(
-        deprecation_reason=(
-            "Use the `deleteMetadata` mutation instead. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-    attribute_update_private_metadata = AttributeUpdatePrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `updatePrivateMetadata` mutation instead. This field will be "
-            "removed after 2020-07-31."
-        )
-    )
-    attribute_clear_private_metadata = AttributeClearPrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `deletePrivateMetadata` mutation instead. This field will be "
-            "removed after 2020-07-31."
-        )
-    )
 
     # attribute value mutations
     attribute_value_create = AttributeValueCreate.Field()