+++ saleor/product/tests/test_product.py
@@ -8,8 +8,8 @@ from prices import Money
 
 from ...account import events as account_events
 from ...attribute.utils import associate_attribute_values_to_instance
+from ...graphql.product.filters import filter_products_by_attributes_values
 from .. import models
-from ..filters import filter_products_by_attributes_values
 from ..models import DigitalContentUrl
 from ..thumbnails import create_product_thumbnails
 from ..utils.costs import get_margin_for_variant