+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable, unreleased changes to this project will be documented in this file.
 
 - Add metadata to shipping zones and shipping methods - by @maarcingebala #6340
 - Add Page Types - #6261 by @IKarbowiak
+- Add metadata to shipping zones and shipping methods - #6340 by @maarcingebala
+- Drop deprecated meta mutations - #6422 by @maarcingebala
 
 ## 2.11.0